proxy: Handle messages with chunked transfer coding
This adds full support for chunked transfer coding,
where appropriate. It is supported for both requests
and responses, but since Elasticsearch does it not
support we're only able to receive chunked requests
but not to forward them. This is actually not a issue
because we need to inspect the body of a request and
thus require it to be completely available in memory.

refs #3013